requirements.txt
<<<<<<< HEAD
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
pgvector==0.2.4
openai==1.3.7
anthropic==0.7.8
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
beautifulsoup4==4.12.2
requests==2.31.0
pypdf2==3.0.1
pytesseract==0.3.10
pillow==10.1.0
open-web-search==1.0.0
=======
# TRA (Talent Resolution Agent) Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Text Processing
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Vector Database
pgvector==0.2.4

# OCR and Document Processing
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0

# API Clients
anthropic==0.7.8
openai==1.3.7

# Data Validation
jsonschema==4.20.0
cerberus==1.3.5

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment Management
python-dotenv==1.0.0

# Async Utilities
asyncio-throttle==1.0.2
aioredis==2.0.1

# File Processing
python-magic==0.4.27
chardet==5.2.0

# Date/Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# Configuration Management
pydantic-settings==2.1.0
typer==0.9.0

# Security
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# Caching
redis==5.0.1

# Task Queue
celery==5.3.4

# Metrics and Analytics
matplotlib==3.8.2
seaborn==0.13.0

# Jupyter Notebook Support (for development)
jupyter==1.0.0
ipykernel==6.26.0
>>>>>>> 56df836 (agent ai)
